Allow pointer unop increment/decrement
This allows us to do `@ptr++` instead of `@ptr+=1`. To allow for this we
need to slightly overhaul the logic as maps were always initialized to
uint regardless of whether they were already set. As increment only used
to work on ints that was ok, but now that it needs to support pointers
it should only be done if there is no default value.