Run memleak tests in CI
This requires to re-build bpftrace with BUILD_ASAN=ON if memleak tests
are required to run.

Currently, the tests are run for all versions of LLVM except for 12 as
that one contains a memory leak.

Also, we disable alloc-dealloc-mismatch errors of ASAN as they produce
false positives in the CI containers for all LLVM versions.